decode -1

+++ src/protobuf/parse/proto.cpp
@@ -370,7 +370,7 @@ namespace proto {
         do {
             const uint8_t next_number = ReadFromBytes<uint8_t>(current, remaining);
             keep_going = (next_number >= 128);
-            result += (next_number & 0x7f) << shift;
+            result += (uint64_t)(next_number & 0x7f) << shift;
             shift += 7;
         } while (keep_going);
         return result;
